+++ src/action.ts
@@ -4,7 +4,7 @@ import GithubReporter from './GithubReporter';
 
 async function main(): Promise<void> {
 	const configFile: string = getInput('config');
-	// Const coverage = Boolean(getInput('coverage'));
+	const coverage = Boolean(getInput('coverage') ?? true);
 
 	const vitest = await startVitest('test', [], {
 		watch: false,
@@ -12,6 +12,11 @@ async function main(): Promise<void> {
 	}, {
 		test: {
 			reporters: [new GithubReporter(), 'default'],
+			coverage: {
+				enabled: coverage,
+				provider: 'custom',
+				customProviderModule: 'github-istanbul-coverage-provider',
+			},
 		},
 	});