+++ src/github-istanbul-coverage-provider.ts
@@ -1,0 +1,2 @@
+import {GithubIstanbulCoverageProviderModule} from '.';
+export default GithubIstanbulCoverageProviderModule;